Move run_command to dedicated module + add support for multiple allowed return code
